requirements.txt
mistralai==1.5.1
openai>=1.0.0
python-dotenv==1.0.1
colorama==0.4.6
requests>=2.31.0
crewai>=0.16.0
crewai-tools>=0.0.15
streamlit>=1.31.0
pathlib>=1.0.1